+++ src/main/java/net/silentchaos512/gear/data/client/ModItemModelProvider.java
@@ -104,6 +104,7 @@ public class ModItemModelProvider extends ItemModelProvider {
         builder(ModItems.AZURE_REPAIR_KIT, itemGenerated);
 
         // Misc
+        builder(ModItems.GUIDE_BOOK, itemGenerated);
         builder(ModItems.BLUEPRINT_PACKAGE, itemGenerated);
         builder(ModItems.FLAX_SEEDS, itemGenerated);
         builder(ModItems.FLUFFY_SEEDS, itemGenerated);