+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- A guide book... Don't get too excited, it only gives a link to the wiki right now
+
 ## [2.6.10] - 2021-03-18
 ### Added
 - Example recipes for generic compounding (metal alloyer, recrystallizer) to JEI [#346]